modules/audio_preprocessing.py
# modules/audio_preprocessing.py
import numpy as np # type: ignore
from typing import Optional
import os
import threading

# NOTE: This is a placeholder for RNNoise Python bindings
# In a real implementation, you would use a proper RNNoise wrapper
# like https://github.com/GregorR/rnnoise-nu or build your own Python bindings
try:
    import rnnoise # type: ignore
    RNNOISE_AVAILABLE = True
except ImportError:
    RNNOISE_AVAILABLE = False
    print("RNNoise not available. Using fallback noise reduction.")

class AudioPreprocessing:
    def __init__(self):
        """Initialize audio preprocessing with noise reduction."""
        self.rnnoise_denoiser = None
        self.is_initialized = False
        self.initialization_lock = threading.Lock()
        
        # Initialize RNNoise in a separate thread
        self.init_thread = threading.Thread(target=self._initialize_rnnoise)
        self.init_thread.daemon = True
        self.init_thread.start()
    
    def _initialize_rnnoise(self):
        """Initialize RNNoise denoiser."""
        with self.initialization_lock:
            if RNNOISE_AVAILABLE:
                try:
                    # This is a placeholder for actual RNNoise initialization
                    # In a real implementation, you would initialize the denoiser
                    self.rnnoise_denoiser = rnnoise.RNNoise()
                    self.is_initialized = True
                    print("RNNoise initialized successfully")
                except Exception as e:
                    print(f"Failed to initialize RNNoise: {str(e)}")
            else:
                # Fallback to simple noise reduction
                print("Using simple noise reduction as fallback")
                self.is_initialized = True
    
    def process_audio(self, audio_data):
        """Process audio with noise reduction."""
        with self.initialization_lock:
            if not self.is_initialized:
                print("Audio preprocessing not yet initialized, returning original audio")
                return audio_data
            
            try:
                if RNNOISE_AVAILABLE and self.rnnoise_denoiser:
                    # This is a placeholder for actual RNNoise processing
                    # In a real implementation, you would call the denoiser
                    # processed_audio = self.rnnoise_denoiser.process_frame(audio_data)
                    processed_audio = audio_data  # Placeholder
                    return processed_audio
                else:
                    # Simple noise reduction using moving average
                    return self._simple_noise_reduction(audio_data)
            except Exception as e:
                print(f"Error processing audio: {str(e)}")
                return audio_data
    
    def _simple_noise_reduction(self, audio_data):
        """Simple noise reduction using spectral subtraction."""
        # Convert to float for processing
        float_data = audio_data.astype(np.float32) / 32768.0
        
        # Simple noise gate
        threshold = 0.01
        gate_mask = np.abs(float_data) > threshold
        gated_audio = float_data * gate_mask
        
        # Convert back to original format
        if isinstance(audio_data, np.ndarray) and audio_data.dtype == np.int16:
            return (gated_audio * 32768.0).astype(np.int16)
        return gated_audio
    
    def cleanup(self):
        """Clean up resources."""
        if RNNOISE_AVAILABLE and self.rnnoise_denoiser:
            # In a real implementation, you would free the resources
            # self.rnnoise_denoiser.destroy()
            self.rnnoise_denoiser = None
